Move the parsing method to GeoPoint
This doesn't really need to have anything to do with JsonObject for most
of its usages
