--- corehq/ex-submodules/couchforms/datatypes.py
@@ -1,10 +1,0 @@
-from collections import namedtuple
-
-
-class GeoPoint(namedtuple('GeoPoint', 'latitude longitude altitude accuracy')):
-    @property
-    def lat_lon(self):
-        return {
-            'lat': self.latitude,
-            'lon': self.longitude
-        }
